requirements.txt
allure-pytest==2.8.18
allure-python-commons==2.8.18
Appium-Python-Client==1.0.2
appnope==0.1.2
argon2-cffi==20.1.0
arrow==0.17.0
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
bleach==3.2.3
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
crypto==1.4.1
cryptography==3.4.7
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
et-xmlfile==1.0.1
Faker==5.0.0
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.1.1
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jdcal==1.4.1
jedi==0.18.0
Jinja2==2.11.2
jsonpath==0.82
jsonschema==3.2.0
MarkupSafe==1.1.1
mistune==0.8.4
Naked==0.1.31
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
nose==1.3.7
notebook==6.2.0
openpyxl==3.0.5
packaging==20.4
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prometheus-client==0.9.0
prompt-toolkit==3.0.14
ptyprocess==0.7.0
py==1.9.0
pycparser==2.20
pycryptodome==3.10.1
Pygments==2.7.4
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.1.1
python-dateutil==2.8.1
PyYAML==5.3.1
pyzmq==22.0.2
qtconsole==5.0.2
QtPy==1.9.0
requests==2.24.0
requests-toolbelt==0.9.1
selenium==3.141.0
Send2Trash==1.5.0
shellescape==3.8.1
six==1.15.0
terminado==0.9.2
testpath==0.4.4
text-unidecode==1.3
toml==0.10.1
tornado==6.1
traitlets==5.0.5
urllib3==1.25.11
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
xmindparser==1.0.8
zipp==3.4.0

threadpool~=1.3.2
